+++ src/crypto.rs
@@ -5,23 +5,23 @@
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
 use aes::Aes256;
-use blake3::Hash;
 use block_modes::block_padding::Pkcs7;
 use block_modes::{BlockMode, Cbc};
 use rand::rngs::OsRng;
 use rand::RngCore;
+use sha2::{Digest, Sha256};
 
 use crate::Error;
 
 type Aes256Cbc = Cbc<Aes256, Pkcs7>;
 
 /// Encrypts data using a password.
-pub(crate) fn encrypt(plaintext: &[u8], pwd: Hash) -> Result<Vec<u8>, Error> {
+pub(crate) fn encrypt(plaintext: &[u8], pwd: &[u8]) -> Result<Vec<u8>, Error> {
     let mut iv = vec![0; 16];
     let mut rng = OsRng::default();
     rng.fill_bytes(&mut iv);
 
-    let cipher = Aes256Cbc::new_from_slices(pwd.as_bytes(), &iv)?;
+    let cipher = Aes256Cbc::new_from_slices(pwd, &iv)?;
     let enc = cipher.encrypt_vec(plaintext);
 
     let ciphertext = iv.into_iter().chain(enc.into_iter()).collect();
@@ -29,14 +29,22 @@ pub(crate) fn encrypt(plaintext: &[u8], pwd: Hash) -> Result<Vec<u8>, Error> {
 }
 
 /// Decrypts data encrypted with `encrypt`.
-pub(crate) fn decrypt(ciphertext: &[u8], pwd: Hash) -> Result<Vec<u8>, Error> {
+pub(crate) fn decrypt(ciphertext: &[u8], pwd: &[u8]) -> Result<Vec<u8>, Error> {
     let iv = &ciphertext[..16];
     let enc = &ciphertext[16..];
 
-    let cipher = Aes256Cbc::new_from_slices(pwd.as_bytes(), iv)?;
+    let cipher = Aes256Cbc::new_from_slices(pwd, iv)?;
     let plaintext = cipher.decrypt_vec(enc)?;
 
     Ok(plaintext)
+}
+
+/// Hashes bytes with sha256 instead of blake3 because it's compatible with web
+pub fn password_hash(bytes: &[u8]) -> Vec<u8> {
+    let mut hasher = Sha256::new();
+    hasher.update(bytes);
+
+    hasher.finalize().to_vec()
 }
 
 #[cfg(test)]
@@ -47,7 +55,9 @@ mod tests {
     fn encrypt_and_decrypt() {
         let seed =
             b"0001020304050607000102030405060700010203040506070001020304050607";
-        let pwd = blake3::hash("greatpassword".as_bytes());
+
+        // hash passwords
+        let pwd: &[u8] = &password_hash(b"greatpassword");
 
         let enc_seed = encrypt(seed, pwd).expect("seed to encrypt ok");
         let enc_seed_t = encrypt(seed, pwd).expect("seed to encrypt ok");